Set up structure for serving analysis
NotificationDtos now have the ability to carry with them analysis
parsed from an on-instrument analysis suite. This is set up as
an array (for multuple attempts) of pipelineRuns, each of which
has some metadata and an array of workflowRuns which make up the
pipeline.

Since it may be the case that sequencing is done but analysis is
not, it is now up to the dto implementation to make the decision
on whether it is fully 'done'.