Introduce DRAGEN parsing with BCLConvert
Use the new analysis parsing infrastructure to parse in
Illumina DRAGEN's BCLConvert workflow. Parsing the required
information requires the presence of many files, so the parsing
is PENDING until all information is found and matches the
samplesheet. If the workflow is still PENDING when Illumina's
'Secondary Analysis Completed' touchfile drops, consider it failed.